+++ CHANGES.rst
@@ -37,6 +37,7 @@ general
 - Adapt :ref:`vendor-hydro2motion` and :ref:`vendor-lst` to infrastructure changes
 - Upgrade libraries Twisted, autobahn, crossbar, msgpack and influxdb
 - Improve Grafana gracefulness when finding a corrupt panel
+- Overhaul configuration subsystem
 
 documentation
 -------------